some quality of life changes
