+++ task_1/examples/4_exec_dir.c
@@ -10,9 +10,9 @@ int main() {
 	pid_t pid;
    	int status, died;
    	switch(pid = fork()) {
-   		case 0: 
-			execl("/bin/ls","ls","-l",NULL); // this is the code the child runs 
-   		default: 
+   		case 0:
+			execl("/bin/ls","ls","-l",NULL); // this is the code the child runs
+   		default:
 			wait(&status);
    	}
 	return 0;
